refactor: Move common constants and type definitions to root
