+++ packages/insight-viewer/src/const.ts
@@ -1,16 +1,4 @@
-import { Viewport } from './types'
-import { handleError } from './utils/common'
-
-export const DefaultViewport: Viewport = {
-  scale: 0,
-  invert: false,
-  hflip: false,
-  vflip: false,
-  x: 0,
-  y: 0,
-  windowWidth: 0,
-  windowCenter: 0,
-}
+import consola from 'consola'
 
 export const LOADER_TYPE = {
   Dicom: 'Dicom',
@@ -18,8 +6,17 @@ export const LOADER_TYPE = {
 } as const
 
 export const CONFIG = {
-  onError: handleError,
+  onError: function handleError(e: Error): void {
+    consola.error(e)
+  },
   Progress: undefined,
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   requestInterceptor: (_request: Request): void => {},
 }
+
+export const LOADING_STATE = {
+  INITIAL: 'initial',
+  LOADING: 'loading',
+  SUCCESS: 'success',
+  FAIL: 'fail',
+} as const
