+++ packages/insight-viewer/src/utils/common/const.ts
@@ -1,0 +1,12 @@
+import { Viewport } from '../../types'
+
+export const DefaultViewport: Viewport = {
+  scale: 0,
+  invert: false,
+  hflip: false,
+  vflip: false,
+  x: 0,
+  y: 0,
+  windowWidth: 0,
+  windowCenter: 0,
+}
